refactor: separate functions into dedicated files based on functionality
